OMAP2 clock: associate clocks with clockdomains at startup
The OMAP2 clock code was missing code to associate clocks with clockdomains
at registration time; fix this.

Resolves "Unhandled fault: alignment exception (0x001) at 0xc02c1b4e" (address
may vary) panic during clock framework init on OMAP2.  The alignment error
was caused by an attempt to dereference a pointer to a string (1-byte aligned)
as if it were a pointer to a structure.

Thanks to Felipe Balbi <felipe.balbi@nokia.com> for originally reporting this
bug.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>