support LED signals on Beagle
Export Beagle's third LED to the LED framework, now that it's
accessible through the twl4030-gpio code.

Claim the VBUS control, defaulting to OFF.  Boards that have an
EHCI connector (not rev B) will need to update this slightly.
(Ideally as part of adding hooks to the ehci-map code supporting
external power switching and overcurrent detection.)

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Tony Lindgren <tony@atomide.com>