omap3 evm: MMC switch support
Update OMAP3 EVM board support to support the MMC/SD switches,
when used with the twl4030 based power card.

Also, less usefully, LEDB as a GPIO LED ... it's on the underside
of the board, so it's probably not very useful to most folk.  The
LCD backlight is controlled by LEDA though; more useful, though
its PWM mode might not be practical on this board (I don't know
how that inductor will affect things).

This is done from schematics, so it needs testing (and maybe some
bugfixes) from folk with an EVM board.

- Dave

Signed-off-by: Tony Lindgren <tony@atomide.com>