omap3 boards list voltage ranges for VMMC1, VMMC2,
It seems the regulator framework is going to require boards
to explicitly configure voltage ranges, instead of letting
them just pick up the entire available range.  So this patch
updates OMAP3 board handling for VMMC1, VMMC2, and VSIM.

Also, change the label for the MMC1 DAT4..DAT7 supply to
"vmmc_aux", so the label isn't specific to MMC1.  Things
like SDIO and eMMC chips may also need a second voltage,
though in that case it's usually 1.8V for chip interface.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Tony Lindgren <tony@atomide.com>