OMAP: McBSP: Always maintain McBSP fclk while active
McBSP fclk must be maintained for the duration of
audio playback or recording. Otherwise the fclk
may get autogated when the PER96M clk is no longer
required by other modules. This results in audio
activity being hang. Also, if the McBSP is run
as a slave, it is possible that words are
randomly missed from the playback. Fix all this
phenomenom by enabling the McBSP fclk
clockactivity bit for the entire active duration
of the McBSP usage.

Signed-off-by: Eero Nurkkala <ext-eero.nurkkala@nokia.com>
Acked-by: Jarkko Nikula <jarkko.nikula@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>