REMOVE OMAP LEGACY CODE: Reset clocks and PM code to mainline
This is to avoid serious merge conflicts when pulling from
mainline.

With tons of the clock patches now in the mainline, it's a good
time to switch over to using the mainline code for clocks and PM.

The remaining clock and pm patches will be merged back in from
Paul's clock branch, and Kevin's pm branch.

Signed-off-by: Tony Lindgren <tony@atomide.com>