Patch: Fix N770 MMC support
Some of the N770's MMC configuration options seem to have been
dropped. This patch adds them back in again.

Note that only the .ocr_mask change was /critical/, but I've added the
.max_freq setting back as well, as the original sources had it. Can
anyone confirm if this is unnecessary?

Secondly, there is support in the original code for a 4wire/higher
speed mode. As I don't have the requisite N770 hardware (I think it
was a rev2 N770?) to test this, I can't really add it back.


commit b67345993d494a1dd5021e29de9611af80cbf1cd
Author: Andrew de Quincey <adq@lidskialf.net>
Date:   Sat May 23 17:43:19 2009 +0100

    Fix n770 MMC support

    Some of the necessary configuration settings for the N770's MMC support had been
    dropped. Add them back in again.

    Signed-off-by: Andrew de Quincey <adq@lidskialf.net>

Signed-off-by: Tony Lindgren <tony@atomide.com>