REMOVE OMAP LEGACY CODE: Remove omap boot tag support
This is not going to mainline as discussed on linux-arm-kernel
list several times earlier. So far it sounds like we should use
platform_data where possible for passing board data to drivers.

Any boot specific options should be passed via kernel cmdline,
which is the Linux standard. If ATAGs are needed, they should be
ARM generic.

Also using device tree could be explored as discussed in the
linux-arm-kernel list.

Signed-off-by: Tony Lindgren <tony@atomide.com>