input: fix rx51 board keymap
The original driver was written with the KEY() macro defined as (col, row)
instead of (row, col) as defined by the matrix keypad infrastructure. So the
keymap was defined accordingly.

Since the driver that was merged upstream uses the matrix keypad
infrastructure, modify the keymap accordingly.

While we are at it, fix the comments in twl4030.h and define PERSISTENT_KEY as
(r,c) instead of (c, r)

Tested on a RX51 (N900) device.

Signed-off-by: Amit Kucheria <amit.kucheria@verdurent.com>
Cc: Tony Lindgren <tony@atomide.com>
Cc: Lauri Leukkunen <lauri.leukkunen@nokia.com>
Cc: Samuel Ortiz <sameo@linux.intel.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>