omap2: Update n8x0 defconfig
Recent DMA API changes broke compile for tusb6010. While
testing the fixes for tusb6010, I had to update the n8x0
defconfig quite a bit. Might as well merge it while at it
to make it more usable.

Also, anybody using n8x0 will most likely want to mount root
on the MMC instead of the onenand to keep the Maemo install
intact. So enable I2C, REGULATOR, MMC to do this. And also
change the root to /dev/mmcblk0p2 instead of the onenand.

Signed-off-by: Tony Lindgren <tony@atomide.com>