Subject: omap: Enable Magic SysRq on serial console ttyOx X-Patchwork-Id: 522001 Message-Id: <1296545441-17070-1-git-send-email-weber@corscience.de> To: linux-omap@vger.kernel.org Cc: Thomas Weber <weber@corscience.de>,  Kevin Hilman <khilman@ti.com> (commit_signer:1/2=50%),  "Govindraj.R" <govindraj.raja@ti.com> (commit_signer:1/2=50%),  Greg Kroah-Hartman <gregkh@suse.de> (commit_signer:1/2=50%),  linux-kernel@vger.kernel.org (open list)
Magic SysRq key is not working for OMAP on new serial
console ttyOx because SUPPORT_SYSRQ is not defined
for omap-serial.

This patch defines SUPPORT_SYSRQ in omap-serial and
enables handling of Magic SysRq character.

Further there is an issue of losing first break character.
Removing the reset of the lsr_break_flag fixes this issue.

Signed-off-by: Thomas Weber <weber@corscience.de>
Acked-by: Govindraj.R <govindraj.raja@ti.com>
Tested-by: Manjunath G Kondaiah <manjugk@ti.com>
Acked-by: Kevin Hilman <khilman@ti.com>