rcu: Avoid build error for third-party modules
The initial definition of __kfree_rcu() checked a static inline function
argument to see if it was a compile-time constant.  Apparently not all
compilers are willing to put up with this at all optimization levels.
Add a nasty comment and remove the warning, relying on the fact that
__kfree_rcu() is called only from kfree_rcu(), which always passes in
a compile-time constant.

Reported-by: Randy Dunlap <randy.dunlap@oracle.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>