intel-iommu: Fix inappropriate use of superpages in __domain_mapping()
After deciding to use a superpage, we were blindly continuing to do so
even when the remaining size to be mapped was not enough for another
superpage.

Fix the logic to switch back to smaller pages as necessary.

This means we can remove the bogus comparisons of nr_pages and sg_res
with lvl_pages — they must *never* be smaller. That would mean we'd set
up a mapping which is longer than we were asked to.

Change those checks to BUG_ON() instead.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>