HACK: iommu/omap: flush page table entries from L2 cache
The functions flush_iopgd_range() and flush_iopte_range()
are not performing any L2 cache operations when flushing
the MMU page table entries. This has resulted in unnecessary
MMU faults on the DSP and IPU processors on OMAP4 (not seen
on OMAP5 though due to the L2 cache differences). Fix this
by adding proper outer cache operations. The asm instruction
used for the inner cache has also been replaced with an
appropriate API.

NOTE:
This is a temporary patch and an up-port of an older product
version patch. The usage of dmac_flush_range() API outside
the dma-mapping layer is very much frowned upon, some usage
existed before within the IOMMU drivers but is no longer
present on 4.9 kernel.

Signed-off-by: Ramesh Gupta <grgupta@ti.com>
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>
[s-anna@ti.com: add commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>