ARM: OMAP: Move omap2_init_common_devices out of init_early
There's no need to call omap2_init_common_devices from init_early.
It reprograms the memory timings for some boards, so it's better to
do it later so we have a chance to get console messages if something
goes wrong.

Move it to happen after omap_serial_init gets called.

Signed-off-by: Tony Lindgren <tony@atomide.com>