arm:omap:am33xx: Update EVM Detection and handling
AM335X EVM comes in 4 flavours. The type of EVM is encoded in the EEPROM
on the Baseboard.

This patch provides api to read this data from eeprom & setup/configure
the evm/devices accordingly. Also, Profile data is retrieved by reading
the CPLD on the daughter board.

In case the eeprom read fails or if thh data is invalid, the fallback
option selects General Purpose EVM with Daughter board connected in
Profile 0.

Also, This patch adds support for BETA EVM.
BETA EVM is set as the default configuration.
Since the peripherals were supported only in profile 0 for
alpha EVM, It makes use of the exisiting GP EVM structure.

Flag gp_evm_revision can be used to differentiate between the
revisions of the GP EVM.

Also, there exist Alpha EVMs with a version string of 1.0B as
well as 1.0A.
Make our revision detection logic check for just 1.0 and rename
defines to match.

Signed-off-by: Patil, Rachna <rachna@ti.com>
Signed-off-by: Tom Rini <trini@ti.com>
Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>
Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Afzal Mohammed <afzal@ti.com>