ARM: OMAP: omap_device: add omap_device_reset to reset all the hwmods in the device
Add a function omap_device_reset() to reset all the hwmods in the hwmod device.
This is intended to be used by device drivers to reset all the hwmods in the
device. This is needed to support some ips like i2c which  may have to do reset
in error paths or in errata.

Signed-off-by: Shubhrajyoti D <shubhrajyoti@ti.com>
Signed-off-by: Paul Walmsley <paul@pwsan.com>