omap: mailbox: Add an API for flushing the FIFO
On some platforms like AM33XX, the mailbox module is a
one-way communication between certain processors.

In these scenarios, the initiator (eg. MPU) is completely
responsible for FIFO maintenance and interrupt generation.
The initiator needs a mechanism to flush the FIFO to avoid
overflows or when going to suspend. Without the FIFO being
flushed, the IP block doesn't get disabled thereby blocking
system suspend.

This patch adds an API which the suspend code can invoke
to flush the FIFO and successfully enter suspend state.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>