arm: am33xx: Add CPSW MII mac select support
This patch fixes missing configuration for MII interface
configuration in device control register. MII mode selection
should be done depending up on the board/phy type.
This was missing in linux, however without this also it was
working as it was configured in uboot. This is required in
linux also in order to remove any dependency on uboot.

Signed-off-by: Chandan Nath <chandan.nath@ti.com>