ARM: OMAP: AM33XX: Mark the essential clocks as ENABLE_ON_INIT
Some of the clocks will never be disabled from the kernel.
To get an accurate use count for such clocks, mark them
with the ENABLE_ON_INIT flag.

Note: Need to check if there's a better way of doing this.
Maybe there should be a clock_get for such clocks.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>