ARM: OMAP: AM33XX: Add EMIF and DDR PHY register offsets and values
When coming out of low power state, the EMIF and DDR
PHY registers need to be restored to the old value.

A naive approach of storing the DDR PHY register values during
suspend and restoring it back during resume will not work
due to a silicon errata related to readback of the DDR PHY
registers.

EMIF registers values can be readback and stored but for
now we assume that the code is running on AM335x EVM with DDR2
which is configured to run @266MHz.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>