ARM: OMAP: AM33XX: Disable self-refresh upon resume
During suspend, we put the SDRAM in self-refresh mode.
If we don't clear this mode in the resume path SDRAM
can automatically enter self-refresh when the CPU is idle.

While this looks to be right behavior, this causes flickers
on the LCD specifically on ICS. So, when coming out of
suspend disable self-refresh mode.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>