ARM: OMAP: AM33XX: Rename the firmware binary
Update the name of the firmware binary to be loaded
onto Cortex-M3 for low power transitions.

If the binary is not provided at the time of kernel
compilation there's a timeout of 60 seconds. To avoid
creating any confusion due to missing binary add a
message indicating that the firmware is now going
to be loaded.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>