ARM: OMAP: AM33XX: Update defconfig with PM support
With most of the PM related code merged in enable
PM support in the defconfig.

Note: PM support depends on the CM3 firmare binary
which is compiled into the kernel image. The latest
image can be obtained from Arago and needs to be placed
in the firmware directory of the kernel.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>