cpufreq: OMAP: Disable cpufreq during suspend
On AM335x, we need to reconfigure the PLLs and voltage
during suspend. With regulator calls in the suspend code
errors were reported by the cpufreq driver in the resume
path. This is probably because cpufreq is suspended
very late in the overall suspend process.

For now, we make use of SUSPEND_PREPARE notification
to block any further OPP changes from the driver.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>