ARM: OMAP: AM33XX: PM: Remove voltage scaling from PM code
Since the cpufreq driver is a consumer of the regulator that
is being used for scaling the voltage during suspend resume,
a regulator_set_voltage() without informing cpufreq gets into
issues with the min_uV and max_uV that both agree upon.

This was leading to suspend getting aborted and system eventually
dying.

A subsequent patch will add back the scaling using the cpufreq
driver itself.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>