ARM: OMAP: AM33XX: CAN: d_can: fix the rx problem
This patch fixes the d_can receiver problem after suspend/resume
cycle.

Details:
After resuming from deepsleep Zero DCAN receiver is not exporting
packets to userspace. This issue mainly because of RAMINIT sequence
is not correct in suspend resume cycle.

This patch replace the tristate open_status flag with two state opened
flag, which is required with the current implementation. Here we end up
with two states one is DCAN is opened or not. If it is opened then we
have to do d_can_stop() and d_can_start() otherwise not required. If
this flag is not there then some unknow bittimming value is programed
to DCAN BTR register which is not expected.

This patch also moves the runtime PM calls from open/close APIs to
DCAN suspend/resume APIs, required along with DCAN RAMINIT and DEINIT

Signed-off-by: AnilKumar Ch <anilkumar@ti.com>